pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.6.2",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]